Add TDD testcases

+++ services/audio_service/client/src/audio_system_manager.cpp
@@ -332,7 +332,7 @@ int32_t AudioSystemManager::GetVolume(AudioVolumeType volumeType) const
             break;
         default:
             AUDIO_ERR_LOG("GetVolume volumeType=%{public}d not supported", volumeType);
-            return (float)ERR_NOT_SUPPORTED;
+            return ERR_NOT_SUPPORTED;
     }
 
     if (volumeType == STREAM_ALL) {
